+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Releases
 v1.23.0-dev (unreleased)
 --------------------
 
--   No changes yet.
+-   Introduces `api/peer.ListImplementation` with `peer/peerlist.List`, a
+    building block that provides peer availability management for peer lists
+    like round-robin, hash-ring.
 
 v1.22.0 (2017-11-14)
 --------------------