+++ CHANGELOG.md
@@ -7,6 +7,8 @@ v1.23.0-dev (unreleased)
 -   Introduces `api/peer.ListImplementation` with `peer/peerlist.List`, a
     building block that provides peer availability management for peer lists
     like round-robin, hash-ring.
+-   Add /x/yarpctest infrastructure to create fake services and requests for
+    tests.
 
 v1.22.0 (2017-11-14)
 --------------------